feet: fixes

+++ src/main/java/com/innowise/orderservice/dto/output/OrderItemOutputDto.java
@@ -1,5 +1,7 @@
 package com.innowise.orderservice.dto.output;
 
+import java.math.BigDecimal;
+
 public record OrderItemOutputDto(
 
         Long id,
@@ -8,9 +10,9 @@ public record OrderItemOutputDto(
 
         String itemName,
 
-        double price,
+        BigDecimal price,
 
-        int quantity
+        Integer quantity
 ) {
