+++ src/main/java/com/innowise/orderservice/dto/output/OrderOutputDto.java
@@ -3,6 +3,7 @@ package com.innowise.orderservice.dto.output;
 import com.innowise.orderservice.model.Status;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -13,7 +14,7 @@ public record OrderOutputDto(
 
         Status status,
 
-        double totalPrice,
+        BigDecimal totalPrice,
 
         boolean deleted,
